Moved MorphologicalProcessor to the namespace and added Masks namespace
