+++ include/MorphologicalProcessor.h
@@ -7,16 +7,21 @@
 #include <unordered_map>
 #include <vector>
 #include <opencv2/opencv.hpp>
+#include <Masks.hpp>
 
 
-class MorphologicalProcessor {
-public:
+using Masks::FieldType;
+using Masks::maskMap;
+using Masks::hmtMaskMap;
+using Masks::hmtComplementMaskMap;
+
+namespace MorphologicalProcessor {
     /**
      * Compute complement (negative) of binary image.
      * @param image Input binary image
      * @return Complemented image
      */
-    static cv::Mat complement(cv::Mat image);
+    cv::Mat complement(cv::Mat image);
 
     /**
      * Compute the union of two binary images.
@@ -24,7 +29,7 @@ public:
      * @param image2 Second input binary image
      * @return Union of input images
      */
-    static cv::Mat imagesUnion(cv::Mat image1, cv::Mat image2);
+    cv::Mat imagesUnion(cv::Mat image1, cv::Mat image2);
 
     /**
      * Check if two images are equal.
@@ -32,19 +37,7 @@ public:
      * @param image2 Second image to compare
      * @return true if images are equal, false otherwise
      */
-    static bool areEqual(cv::Mat image1, cv::Mat image2);
-
-    enum class FieldType {
-        WHITE,
-        BLACK,
-        WHITE_MARKER,
-        BLACK_MARKER,
-        INACTIVE
-    };
-
-    static const std::unordered_map<int, std::vector<std::vector<FieldType> > > maskMap;
-    static const std::unordered_map<int, std::vector<std::vector<FieldType> > > hmtMaskMap;
-    static const std::unordered_map<int, std::vector<std::vector<FieldType> > > hmtComplementMaskMap;
+    bool areEqual(cv::Mat image1, cv::Mat image2);
 
     /**
      * Apply dilation morphological operation.
@@ -53,7 +46,7 @@ public:
      * @param maskMapping Map of available structural elements
      * @return Dilated image
      */
-    static cv::Mat dilation(cv::Mat image, int maskNumber,
+    cv::Mat dilation(cv::Mat image, int maskNumber,
                      const std::unordered_map<int, std::vector<std::vector<FieldType> > > &maskMapping = maskMap);
 
     /**
@@ -63,7 +56,7 @@ public:
      * @param maskMapping Map of available structural elements
      * @return Eroded image
      */
-    static cv::Mat erosion(cv::Mat image, int maskNumber,
+    cv::Mat erosion(cv::Mat image, int maskNumber,
                     const std::unordered_map<int, std::vector<std::vector<FieldType> > > &maskMapping = maskMap);
 
     /**
@@ -73,7 +66,7 @@ public:
      * @param maskMapping Map of available structural elements
      * @return Opened image
      */
-    static cv::Mat opening(const cv::Mat &image, int maskNumber,
+    cv::Mat opening(const cv::Mat &image, int maskNumber,
                     const std::unordered_map<int, std::vector<std::vector<FieldType> > > &maskMapping = maskMap);
 
     /**
@@ -83,7 +76,7 @@ public:
      * @param maskMapping Map of available structural elements
      * @return Closed image
      */
-    static cv::Mat closing(const cv::Mat &image, int maskNumber,
+    cv::Mat closing(const cv::Mat &image, int maskNumber,
                     const std::unordered_map<int, std::vector<std::vector<FieldType> > > &maskMapping = maskMap);
 
     /**
@@ -93,23 +86,9 @@ public:
      * @param maskMapping Map of available HMT masks
      * @return Transformed image
      */
-    static cv::Mat hmt(const cv::Mat &image, int maskNumber,
-                       const std::unordered_map<int, std::vector<std::vector<FieldType>>> &maskMapping = hmtMaskMap);
-
-    /**
-     * Create a color map for visualization.
-     * @return Vector of BGR colors
-     */
-    static std::vector<cv::Vec3b> createColorMap();
-
-    /**
-     * Apply a color map to grayscale image.
-     * @param grayscaleMask Input grayscale image
-     * @param colorMap Color map to apply
-     * @return Colored image
-     */
-    static cv::Mat applyColorMap(const cv::Mat &grayscaleMask, const std::vector<cv::Vec3b> &colorMap);
-};
+    cv::Mat hmt(const cv::Mat &image, int maskNumber,
+                const std::unordered_map<int, std::vector<std::vector<FieldType> > > &maskMapping = hmtMaskMap);
+}
 
 
 #endif //MORPHOLOGICALPROCESSOR_H
